Hice una primera version algo estable ya anda el juego aunque todavia toma las figuras como rectangulos!! despues lo vamos puliendo!! pruebenlo

+++ TPTaller/TP3/src/Direccion.cpp
@@ -19,11 +19,11 @@ Direccion::~Direccion() {
 }
 
 
-float Direccion::getFi(){
+double Direccion::getFi(){
 	return fi;
 }
 
-void Direccion::setFi(float fi)
+void Direccion::setFi(double fi)
 {
 	this->fi = fi;
 }
